fix: Resolve login failure by correctly registering dashboard routes
This commit fixes a critical bug that prevented users from logging in. The issue was caused by the new dashboard routes not being correctly registered in the main application file (`app.js`).

This commit introduces a new `routes/dashboard.js` file to handle all dashboard-related routes and registers it in `app.js` with the `/dashboard` prefix. This ensures that the application can correctly handle requests to the new role-specific dashboard URLs.
