+++ routes/police.js
@@ -4,75 +4,6 @@ const { PrismaClient } = require('@prisma/client');
 const prisma = new PrismaClient();
 const { checkRole } = require('../middleware/auth');
 
-router.get('/dashboard/police', checkRole(['Police']), async (req, res) => {
-  const user = await prisma.user.findUnique({
-    where: { id: req.session.userId },
-    include: { role: true },
-  });
-
-  const bookingsToday = await prisma.booking.count({
-    where: {
-      bookingDate: {
-        gte: new Date(new Date().setHours(0, 0, 0, 0)),
-      },
-    },
-  });
-
-  const overdueHolds = await prisma.booking.count({
-    where: {
-      bookingDate: {
-        lt: new Date(Date.now() - 24 * 60 * 60 * 1000),
-      },
-      status: 'New Arrest',
-    },
-  });
-
-  const pendingSubmissions = await prisma.booking.count({
-    where: {
-      case: {
-        status: 'New Arrest',
-      },
-    },
-  });
-
-  const inCustody = await prisma.booking.count({
-    where: {
-      status: 'In-Custody',
-    },
-  });
-
-  const recentActivity = await prisma.actionHistory.findMany({
-    where: {
-      user: {
-        role: {
-          name: 'Police',
-        },
-      },
-    },
-    orderBy: {
-      timestamp: 'desc',
-    },
-    take: 10,
-    include: {
-      case: true,
-      user: true,
-    },
-  });
-
-  const breadcrumbs = [
-    { name: 'Dashboard', url: '/dashboard/police' }
-  ];
-  res.render('police/dashboard', {
-    user,
-    bookingsToday,
-    overdueHolds,
-    pendingSubmissions,
-    inCustody,
-    recentActivity,
-    page: '/dashboard/police',
-    breadcrumbs
-  });
-});
 
 router.get('/people-redesigned', checkRole(['Police']), async (req, res) => {
   const user = await prisma.user.findUnique({
