+++ routes/prosecutor.js
@@ -23,67 +23,5 @@ router.post('/cases/:caseId/notes', checkRole(['Prosecutor', 'Court']), async (r
   }
 });
 
-router.get('/dashboard/prosecutor', checkRole(['Prosecutor']), async (req, res) => {
-  const user = await prisma.user.findUnique({
-    where: { id: req.session.userId },
-    include: { role: true },
-  });
-
-  const cases = await prisma.case.findMany({
-    where: {
-      actions: {
-        some: {
-          userId: user.id,
-        },
-      },
-    },
-    include: {
-      booking: {
-        include: {
-          person: true,
-        },
-      },
-    },
-  });
-
-  const remandRequests = await prisma.remandRequest.findMany({
-    where: { status: 'pending' },
-    include: {
-      booking: {
-        include: {
-          person: true,
-          case: true,
-        },
-      },
-    },
-  });
-
-  const upcomingHearings = await prisma.hearing.findMany({
-    where: {
-      case: {
-        actions: {
-          some: {
-            userId: user.id,
-          },
-        },
-      },
-      hearingDate: {
-        gte: new Date(),
-      },
-    },
-  });
-
-  const breadcrumbs = [
-    { name: 'Dashboard', url: '/dashboard/prosecutor' }
-  ];
-  res.render('prosecutor/dashboard', {
-    user,
-    cases,
-    remandRequests,
-    upcomingHearings,
-    page: '/prosecutor/dashboard',
-    breadcrumbs
-  });
-});
 
 module.exports = router;
